+++ cmd/image.go
@@ -251,6 +251,8 @@ func init() {
 	ImageCmd.AddCommand(imageStatusCmd)
 	ImageCmd.AddCommand(imageSetMaxSessionCmd)
 	ImageCmd.AddCommand(imageWarmupStatusCmd)
+	ImageCmd.AddCommand(imageSetPreOpenCmd)
+	ImageCmd.AddCommand(imageDescribePreOpenCmd)
 }
 
 func runImageCreate(cmd *cobra.Command, args []string) error {