+++ cmd/image_list_helper_test.go
@@ -171,6 +171,10 @@ func (m *mockImageListClient) DescribeWarmUpStatusOpen(ctx context.Context, requ
 	return nil, fmt.Errorf("not implemented")
 }
 
+func (m *mockImageListClient) DescribeImageReserveMinAmount(ctx context.Context, request *client.DescribeImageReserveMinAmountRequest) (*client.DescribeImageReserveMinAmountResponse, error) {
+	return nil, fmt.Errorf("not implemented")
+}
+
 func (m *mockImageListClient) ListTag(ctx context.Context) (*client.ListTagResponse, error) {
 	return nil, fmt.Errorf("not implemented")
 }
@@ -200,6 +204,10 @@ func (m *mockImageListClient) ListSharedDockerRepos(ctx context.Context, request
 }
 
 func (m *mockImageListClient) CreateSimpleOfficeSite(ctx context.Context, request *client.CreateSimpleOfficeSiteRequest) (*client.CreateSimpleOfficeSiteResponse, error) {
+	return nil, fmt.Errorf("not implemented")
+}
+
+func (m *mockImageListClient) UpdateImageReserveMinAmount(ctx context.Context, request *client.UpdateImageReserveMinAmountRequest) (*client.UpdateImageReserveMinAmountResponse, error) {
 	return nil, fmt.Errorf("not implemented")
 }